+++ src/app/account/signin/mfa.tsx
@@ -2,7 +2,7 @@ import { Panel } from '@/components/panel'
 import { Button, Flex, Loader, Text, TextInput } from '@mantine/core'
 import { isNotEmpty, useForm } from '@mantine/form'
 import { useSignIn, useUser } from '@clerk/nextjs'
-import { useMutation } from '@/components/common'
+import { useMutation } from '@/common'
 import type { SignInResource } from '@clerk/types'
 import type { MFAState } from './logic'
 import { errorToString } from '@/lib/errors'
