+++ src/app/admin/safeinsights/page.tsx
@@ -1,4 +1,5 @@
 import { OrgsAdminTable } from './table'
+// eslint-disable-next-line no-restricted-imports
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query'
 import { fetchOrgsStatsAction } from '@/server/actions/org.actions'
