+++ src/app/admin/safeinsights/table.tsx
@@ -5,7 +5,7 @@ import * as R from 'remeda'
 import { FC, useMemo, useState } from 'react'
 import { deleteOrgAction, fetchOrgsStatsAction } from '@/server/actions/org.actions'
 import { getNewOrg } from '@/schema/org'
-import { useMutation, useQuery, useQueryClient } from '@/components/common'
+import { useMutation, useQuery, useQueryClient } from '@/common'
 import { PencilIcon, TrashIcon, UsersIcon } from '@phosphor-icons/react/dist/ssr'
 import { ActionIcon, Box, Button, Flex, Group, Modal, Title, Tooltip } from '@mantine/core'
 import { SuretyGuard } from '@/components/surety-guard'
