+++ src/app/admin/team/[orgSlug]/settings/organization-settings-edit.tsx
@@ -2,7 +2,7 @@
 
 import { Stack, TextInput, Textarea, Grid, Text, Flex, Title, Button, Group, Divider } from '@mantine/core'
 import { useForm, type UseFormReturnType } from '@mantine/form'
-import { zodResolver, useMutation } from '@/components/common'
+import { zodResolver, useMutation } from '@/common'
 import { notifications } from '@mantine/notifications'
 import { FormFieldLabel } from '@/components/form-field-label'
 import { z } from 'zod'
