+++ src/app/reviewer/[orgSlug]/study/[studyId]/review/decrypt-results.tsx
@@ -9,8 +9,7 @@ import { Button, Group, Stack, Textarea } from '@mantine/core'
 import { isNotEmpty, useForm } from '@mantine/form'
 import { notifications } from '@mantine/notifications'
 import * as Sentry from '@sentry/nextjs'
-import { useQuery } from '@/components/common'
-import { useMutation } from '@tanstack/react-query'
+import { useMutation, useQuery } from '@/common'
 import { useParams } from 'next/navigation'
 import { FC, useState } from 'react'
 import { ResultsReader } from 'si-encryption/job-results/reader'
