+++ src/hooks/query-wrappers.ts
@@ -1,12 +1,18 @@
+// eslint-disable-next-line no-restricted-imports
 import {
-    useQuery,
-    useMutation,
+    useQuery as useTanStackQuery,
+    useMutation as useTanStackMutation,
     type UseQueryOptions,
     type UseMutationOptions,
     type UseQueryResult,
     type UseMutationResult,
+    useQueryClient,
+    skipToken,
 } from '@tanstack/react-query'
+
 import { type ActionResponse, isActionError, ActionFailure } from '@/lib/errors'
+
+export { useTanStackMutation, useTanStackQuery, useQueryClient, skipToken }
 
 // Helper function to process response - handles ActionResponse format and raw responses
 function processResponse<T>(response: ActionResponse<T>): T {
@@ -20,13 +26,13 @@ function processResponse<T>(response: ActionResponse<T>): T {
 }
 
 // Wrapped useQuery that automatically handles error responses
-export function useWrappedQuery<TApiData>(
+export function useQuery<TApiData>(
     options: {
         queryKey: readonly unknown[]
         queryFn: () => Promise<ActionResponse<TApiData>>
     } & Omit<UseQueryOptions<TApiData, Error, TApiData>, 'queryFn' | 'queryKey'>,
 ): UseQueryResult<TApiData, Error> {
-    return useQuery<TApiData, Error>({
+    return useTanStackQuery<TApiData, Error>({
         ...options,
         queryFn: async () => {
             const response = await options.queryFn()
@@ -36,12 +42,12 @@ export function useWrappedQuery<TApiData>(
 }
 
 // Wrapped useMutation that automatically handles error responses
-export function useWrappedMutation<TApiData, TVariables = void>(
+export function useMutation<TApiData, TVariables = void>(
     options: {
         mutationFn: (variables: TVariables) => Promise<ActionResponse<TApiData>>
     } & Omit<UseMutationOptions<TApiData, Error, TVariables>, 'mutationFn'>,
 ): UseMutationResult<TApiData, Error, TVariables> {
-    return useMutation<TApiData, Error, TVariables>({
+    return useTanStackMutation<TApiData, Error, TVariables>({
         ...options,
         mutationFn: async (variables: TVariables) => {
             const response = await options.mutationFn(variables)
